src/requirements.txt
click==8.1.2
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
msgpack==1.0.3
packaging==21.3
pyparsing==3.0.8
python-can==4.0.0
PyYAML==6.0
typing_extensions==4.2.0
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
aniso8601==9.0.1
Flask-RESTful==0.3.9
pytz==2022.1
six==1.16.0
